requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
lpips>=0.1.4
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=5.4.0
